+++ src/gicv3.rs
@@ -3,23 +3,20 @@
 
 //! Driver for the Arm Generic Interrupt Controller version 3 (or 4).
 
+#[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
+pub mod cpu_interface;
 pub mod redistributor;
 pub mod registers;
 
-use crate::gicv3::redistributor::GicRedistributor;
-use crate::gicv3::registers::{Gicd, GicdCtlr, GicrSgi, GicrTyper};
 #[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
-use crate::sysreg::{
-    read_icc_hppir0_el1, read_icc_hppir1_el1, read_icc_iar0_el1, read_icc_iar1_el1,
-    write_icc_asgi1r_el1, write_icc_ctlr_el1, write_icc_eoir0_el1, write_icc_eoir1_el1,
-    write_icc_igrpen0_el1, write_icc_igrpen1_el1, write_icc_pmr_el1, write_icc_sgi0r_el1,
-    write_icc_sgi1r_el1, write_icc_sre_el1, IccSre
-};
+use crate::sysreg::write_icc_ctlr_el1;
 use crate::{IntId, Trigger};
 use core::ptr::NonNull;
-use registers::Typer;
-use safe_mmio::fields::ReadPureWrite;
-use safe_mmio::{UniqueMmioPointer, field, field_shared};
+#[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
+use cpu_interface::GicCpuInterface;
+use redistributor::GicRedistributor;
+use registers::{Gicd, GicdCtlr, GicrSgi, GicrTyper, Typer};
+use safe_mmio::{UniqueMmioPointer, field, field_shared, fields::ReadPureWrite};
 use thiserror::Error;
 
 /// An error which may be returned from operations on a GIC Redistributor.
@@ -136,7 +133,7 @@ impl GicV3<'_> {
     #[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
     pub fn init_cpu(&mut self, cpu: usize) {
         // Enable system register access.
-        write_icc_sre_el1(IccSre::SRE);
+        GicCpuInterface::enable_system_register_el1(true);
 
         // Ignore error in case core is already awake.
         let _ = self.redistributor_mark_core_awake(cpu);
@@ -177,19 +174,7 @@ impl GicV3<'_> {
         }
 
         // Enable group 1 for the current security state.
-        Self::enable_group1(true);
-    }
-
-    /// Enables or disables group 0 interrupts.
-    #[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
-    pub fn enable_group0(enable: bool) {
-        write_icc_igrpen0_el1(if enable { 0x01 } else { 0x00 });
-    }
-
-    /// Enables or disables group 1 interrupts for the current security state.
-    #[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
-    pub fn enable_group1(enable: bool) {
-        write_icc_igrpen1_el1(if enable { 0x01 } else { 0x00 });
+        GicCpuInterface::enable_group1(true);
     }
 
     /// Enables or disables the interrupt with the given ID.
@@ -233,14 +218,6 @@ impl GicV3<'_> {
         for cpu in 0..self.cpu_count {
             self.redistributor(cpu).enable_all_interrupts(enable);
         }
-    }
-
-    /// Sets the priority mask for the current CPU core.
-    ///
-    /// Only interrupts with a higher priority (numerically lower) will be signalled.
-    #[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
-    pub fn set_priority_mask(min_priority: u8) {
-        write_icc_pmr_el1(min_priority.into());
     }
 
     /// Sets the priority of the interrupt with the given ID.
@@ -295,85 +272,6 @@ impl GicV3<'_> {
             set_bit(field!(self.gicd, igroupr).into(), intid.0 as usize);
             clear_bit(field!(self.gicd, igrpmodr).into(), intid.0 as usize);
         };
-    }
-
-    /// Sends a group `group` software-generated interrupt (SGI) to the given cores.
-    #[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
-    pub fn send_sgi(intid: IntId, target: SgiTarget, group: SgiTargetGroup) {
-        assert!(intid.is_sgi());
-
-        let sgi_value = match target {
-            SgiTarget::All => {
-                let irm = 0b1;
-                (u64::from(intid.0 & 0x0f) << 24) | (irm << 40)
-            }
-            SgiTarget::List {
-                affinity3,
-                affinity2,
-                affinity1,
-                target_list,
-            } => {
-                let irm = 0b0;
-                u64::from(target_list)
-                    | (u64::from(affinity1) << 16)
-                    | (u64::from(intid.0 & 0x0f) << 24)
-                    | (u64::from(affinity2) << 32)
-                    | (irm << 40)
-                    | (u64::from(affinity3) << 48)
-            }
-        };
-
-        match group {
-            SgiTargetGroup::Group0 => write_icc_sgi0r_el1(sgi_value),
-            SgiTargetGroup::CurrentGroup1 => write_icc_sgi1r_el1(sgi_value),
-            SgiTargetGroup::OtherGroup1 => write_icc_asgi1r_el1(sgi_value),
-        }
-    }
-
-    /// Gets the ID of the highest priority pending group `group` interrupt on the CPU interface.
-    ///
-    /// Returns `None` if there is no pending interrupt of sufficient priority.
-    #[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
-    pub fn get_pending_interrupt(group: InterruptGroup) -> Option<IntId> {
-        let icc_hppir = match group {
-            InterruptGroup::Group0 => read_icc_hppir0_el1(),
-            InterruptGroup::Group1 => read_icc_hppir1_el1(),
-        };
-
-        let intid = IntId(icc_hppir);
-        if intid == IntId::SPECIAL_NONE {
-            None
-        } else {
-            Some(intid)
-        }
-    }
-
-    /// Gets the ID of the highest priority signalled group `group` interrupt, and acknowledges it.
-    ///
-    /// Returns `None` if there is no pending interrupt of sufficient priority.
-    #[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
-    pub fn get_and_acknowledge_interrupt(group: InterruptGroup) -> Option<IntId> {
-        let icc_iar = match group {
-            InterruptGroup::Group0 => read_icc_iar0_el1(),
-            InterruptGroup::Group1 => read_icc_iar1_el1(),
-        };
-
-        let intid = IntId(icc_iar);
-        if intid == IntId::SPECIAL_NONE {
-            None
-        } else {
-            Some(intid)
-        }
-    }
-
-    /// Informs the interrupt controller that the CPU has completed processing the given group `group` interrupt.
-    /// This drops the interrupt priority and deactivates the interrupt.
-    #[cfg(any(test, feature = "fakes", target_arch = "aarch64", target_arch = "arm"))]
-    pub fn end_interrupt(intid: IntId, group: InterruptGroup) {
-        match group {
-            InterruptGroup::Group0 => write_icc_eoir0_el1(intid.0),
-            InterruptGroup::Group1 => write_icc_eoir1_el1(intid.0),
-        }
     }
 
     /// Returns information about what the GIC implementation supports.